PHP client with BSON-encoded proto3 over GoRPC+

+++ test.go
@@ -25,17 +25,24 @@ package main
 // This Go script shouldn't rely on any packages that aren't in the standard
 // library, since that would require the user to bootstrap before running it.
 import (
+	"bytes"
 	"encoding/json"
+	"errors"
 	"flag"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"log"
+	"net/http"
+	"net/url"
 	"os"
 	"os/exec"
 	"os/signal"
 	"path"
 	"sort"
+	"strconv"
+	"strings"
+	"sync"
 	"syscall"
 	"time"
 )
@@ -56,8 +63,23 @@ var (
 	logPass  = flag.Bool("log-pass", false, "log test output even if it passes")
 	timeout  = flag.Duration("timeout", 10*time.Minute, "timeout for each test")
 	pull     = flag.Bool("pull", true, "re-pull the bootstrap image, in case it's been updated")
+	docker   = flag.Bool("docker", true, "run tests with Docker")
+	shard    = flag.Int("shard", -1, "if N>=0, run the tests whose Shard field matches N")
+	reshard  = flag.Int("rebalance", 0, "if N>0, check the stats and group tests into N similarly-sized bins by average run time")
+	keepData = flag.Bool("keep-data", false, "don't delete the per-test VTDATAROOT subfolders")
+	printLog = flag.Bool("print-log", false, "print the log of each failed test (or all tests if -log-pass) to the console")
+	follow   = flag.Bool("follow", false, "print test output as it runs, instead of waiting to see if it passes or fails")
+
+	remoteStats = flag.String("remote-stats", "", "url to send remote stats")
 
 	extraArgs = flag.String("extra-args", "", "extra args to pass to each test")
+)
+
+var vtDataRoot = os.Getenv("VTDATAROOT")
+
+const (
+	statsFileName  = "test/stats.json"
+	configFileName = "test/config.json"
 )
 
 // Config is the overall object serialized in test/config.json.
@@ -67,27 +89,52 @@ type Config struct {
 
 // Test is an entry from the test/config.json file.
 type Test struct {
-	Name, File, Args, Command string
+	File, Args, Command string
 
+	// Manual means it won't be run unless explicitly specified.
+	Manual bool
+
+	// Shard is used to split tests among workers.
+	Shard int
+
+	name     string
 	cmd      *exec.Cmd
 	runIndex int
+
+	pass, fail int
 }
 
 // run executes a single try.
 // dir is the location of the vitess repo to use.
+// dataDir is the VTDATAROOT to use for this run.
 // returns the combined stdout+stderr and error.
-func (t *Test) run(dir string) ([]byte, error) {
+func (t *Test) run(dir, dataDir string) ([]byte, error) {
 	testCmd := t.Command
 	if testCmd == "" {
-		// Teardown is unnecessary since Docker kills everything.
-		testCmd = fmt.Sprintf("make build && test/%s -v --skip-teardown %s", t.File, t.Args)
+		testCmd = fmt.Sprintf("make build && test/%s -v --skip-build --keep-logs %s", t.File, t.Args)
+		if *docker {
+			// Teardown is unnecessary since Docker kills everything.
+			testCmd += " --skip-teardown"
+		}
 		if *extraArgs != "" {
 			testCmd += " " + *extraArgs
 		}
 	}
-	dockerCmd := exec.Command(path.Join(dir, "docker/test/run.sh"), *flavor, testCmd)
-	dockerCmd.Dir = dir
-	t.cmd = dockerCmd
+
+	if *docker {
+		t.cmd = exec.Command(path.Join(dir, "docker/test/run.sh"), *flavor, testCmd)
+	} else {
+		t.cmd = exec.Command("bash", "-c", testCmd)
+	}
+	t.cmd.Dir = dir
+
+	// Put everything in a unique dir, so we can copy and/or safely delete it.
+	// Also try to make them use different port ranges
+	// to mitigate failures due to zombie processes.
+	t.cmd.Env = updateEnv(os.Environ(), map[string]string{
+		"VTDATAROOT":  dataDir,
+		"VTPORTSTART": strconv.FormatInt(int64(getPortStart(100)), 10),
+	})
 
 	// Stop the test if it takes too long.
 	done := make(chan struct{})
@@ -98,15 +145,30 @@ func (t *Test) run(dir string) ([]byte, error) {
 		case <-done:
 		case <-timer.C:
 			t.logf("timeout exceeded")
-			if dockerCmd.Process != nil {
-				dockerCmd.Process.Signal(syscall.SIGTERM)
+			if t.cmd.Process != nil {
+				t.cmd.Process.Signal(syscall.SIGTERM)
 			}
 		}
 	}()
 
-	// Run the test.
 	defer close(done)
-	return dockerCmd.CombinedOutput()
+
+	// Capture test output.
+	buf := &bytes.Buffer{}
+	t.cmd.Stdout = buf
+	if *follow {
+		t.cmd.Stdout = io.MultiWriter(t.cmd.Stdout, os.Stdout)
+	}
+	t.cmd.Stderr = t.cmd.Stdout
+
+	// Run the test.
+	err := t.cmd.Run()
+	if err == nil {
+		t.pass++
+	} else {
+		t.fail++
+	}
+	return buf.Bytes(), err
 }
 
 // stop will terminate the test if it's running.
@@ -121,9 +183,9 @@ func (t *Test) stop() {
 
 func (t *Test) logf(format string, v ...interface{}) {
 	if *runCount > 1 {
-		log.Printf("%v[%v/%v]: %v", t.Name, t.runIndex+1, *runCount, fmt.Sprintf(format, v...))
+		log.Printf("%v[%v/%v]: %v", t.name, t.runIndex+1, *runCount, fmt.Sprintf(format, v...))
 	} else {
-		log.Printf("%v: %v", t.Name, fmt.Sprintf(format, v...))
+		log.Printf("%v: %v", t.name, fmt.Sprintf(format, v...))
 	}
 }
 
@@ -150,7 +212,7 @@ func main() {
 	log.Printf("Output directory: %v", outDir)
 
 	// Get test configs.
-	configData, err := ioutil.ReadFile("test/config.json")
+	configData, err := ioutil.ReadFile(configFileName)
 	if err != nil {
 		log.Fatalf("Can't read config file: %v", err)
 	}
@@ -158,44 +220,45 @@ func main() {
 	if err := json.Unmarshal(configData, &config); err != nil {
 		log.Fatalf("Can't parse config file: %v", err)
 	}
-	log.Printf("Bootstrap flavor: %v", *flavor)
 
-	// Re-pull image.
-	if *pull {
-		image := "vitess/bootstrap:" + *flavor
-		pullTime := time.Now()
-		log.Printf("Pulling %v...", image)
-		cmd := exec.Command("docker", "pull", image)
-		if out, err := cmd.CombinedOutput(); err != nil {
-			log.Fatalf("Can't pull image: %v\n%s", err, out)
+	// Resharding.
+	if *reshard > 0 {
+		if err := reshardTests(&config, *reshard); err != nil {
+			log.Fatalf("resharding error: %v", err)
 		}
-		log.Printf("Image pulled in %v", time.Since(pullTime))
+		log.Printf("Saving updated config...")
+		data, err := json.MarshalIndent(config, "", "\t")
+		if err != nil {
+			log.Fatalf("can't save new config: %v", err)
+		}
+		if err := ioutil.WriteFile(configFileName, data, 0644); err != nil {
+			log.Fatalf("can't write new config: %v", err)
+		}
+		return
 	}
 
-	// Positional args specify which tests to run.
-	// If none specified, run all tests in alphabetical order.
-	var tests []*Test
-	if flag.NArg() > 0 {
-		for _, name := range flag.Args() {
-			t, ok := config.Tests[name]
-			if !ok {
-				log.Fatalf("Unknown test: %v", name)
+	if *docker {
+		log.Printf("Bootstrap flavor: %v", *flavor)
+
+		// Re-pull image.
+		if *pull {
+			image := "vitess/bootstrap:" + *flavor
+			pullTime := time.Now()
+			log.Printf("Pulling %v...", image)
+			cmd := exec.Command("docker", "pull", image)
+			if out, err := cmd.CombinedOutput(); err != nil {
+				log.Fatalf("Can't pull image: %v\n%s", err, out)
 			}
-			t.Name = name
-			tests = append(tests, t)
+			log.Printf("Image pulled in %v", time.Since(pullTime))
 		}
 	} else {
-		var names []string
-		for name := range config.Tests {
-			names = append(names, name)
-		}
-		sort.Strings(names)
-		for _, name := range names {
-			t := config.Tests[name]
-			t.Name = name
-			tests = append(tests, t)
+		if vtDataRoot == "" {
+			log.Fatalf("VTDATAROOT env var must be set in -docker=false mode. Make sure to source dev.env.")
 		}
 	}
+
+	// Pick the tests to run.
+	tests := selectedTests(&config)
 
 	// Duplicate tests.
 	if *runCount > 1 {
@@ -211,18 +274,24 @@ func main() {
 		tests = dup
 	}
 
-	// Copy working repo to tmpDir.
-	tmpDir, err := ioutil.TempDir(os.TempDir(), "vt_")
-	if err != nil {
-		log.Fatalf("Can't create temp dir in %v", os.TempDir())
-	}
-	log.Printf("Copying working repo to temp dir %v", tmpDir)
-	if out, err := exec.Command("cp", "-R", ".", tmpDir).CombinedOutput(); err != nil {
-		log.Fatalf("Can't copy working repo to temp dir %v: %v: %s", tmpDir, err, out)
-	}
-	// The temp copy needs permissive access so the Docker user can read it.
-	if out, err := exec.Command("chmod", "-R", "go=u", tmpDir).CombinedOutput(); err != nil {
-		log.Printf("Can't set permissions on temp dir %v: %v: %s", tmpDir, err, out)
+	vtTop := "."
+	tmpDir := ""
+	if *docker {
+		// Copy working repo to tmpDir.
+		// This doesn't work outside Docker since it messes up GOROOT.
+		tmpDir, err = ioutil.TempDir(os.TempDir(), "vt_")
+		if err != nil {
+			log.Fatalf("Can't create temp dir in %v", os.TempDir())
+		}
+		log.Printf("Copying working repo to temp dir %v", tmpDir)
+		if out, err := exec.Command("cp", "-R", ".", tmpDir).CombinedOutput(); err != nil {
+			log.Fatalf("Can't copy working repo to temp dir %v: %v: %s", tmpDir, err, out)
+		}
+		// The temp copy needs permissive access so the Docker user can read it.
+		if out, err := exec.Command("chmod", "-R", "go=u", tmpDir).CombinedOutput(); err != nil {
+			log.Printf("Can't set permissions on temp dir %v: %v: %s", tmpDir, err, out)
+		}
+		vtTop = tmpDir
 	}
 
 	// Keep stats.
@@ -260,32 +329,57 @@ func main() {
 				}
 
 				test.logf("running (try %v/%v)...", try, *retryMax)
-				start := time.Now()
-				output, err := test.run(tmpDir)
 
-				// Save test output.
+				// Make a unique VTDATAROOT.
+				dataDir, err := ioutil.TempDir(vtDataRoot, "vt_")
+				if err != nil {
+					test.logf("Failed to create temporary subdir in VTDATAROOT: %v", vtDataRoot)
+					failed++
+					break
+				}
+
+				// Run the test.
+				start := time.Now()
+				output, err := test.run(vtTop, dataDir)
+				duration := time.Since(start)
+
+				// Save/print test output.
 				if err != nil || *logPass {
-					outFile := fmt.Sprintf("%v-%v.%v.log", test.Name, test.runIndex+1, try)
+					if *printLog && !*follow {
+						test.logf("%s\n", output)
+					}
+					outFile := fmt.Sprintf("%v-%v.%v.log", test.name, test.runIndex+1, try)
 					test.logf("saving test output to %v", outFile)
 					if fileErr := ioutil.WriteFile(path.Join(outDir, outFile), output, os.FileMode(0644)); fileErr != nil {
 						test.logf("WriteFile error: %v", fileErr)
 					}
 				}
 
+				// Clean up the unique VTDATAROOT.
+				if !*keepData {
+					if err := os.RemoveAll(dataDir); err != nil {
+						test.logf("WARNING: can't remove temporary VTDATAROOT: ", err)
+					}
+				}
+
 				if err != nil {
 					// This try failed.
-					test.logf("FAILED (try %v/%v) in %v: %v", try, *retryMax, time.Since(start), err)
+					test.logf("FAILED (try %v/%v) in %v: %v", try, *retryMax, duration, err)
+					testFailed(test.name)
 					continue
 				}
 
+				testPassed(test.name, duration)
+
 				if try == 1 {
 					// Passed on the first try.
-					test.logf("PASSED in %v", time.Since(start))
+					test.logf("PASSED in %v", duration)
 					passed++
 				} else {
 					// Passed, but not on the first try.
-					test.logf("FLAKY (1/%v passed in %v)", try, time.Since(start))
+					test.logf("FLAKY (1/%v passed in %v)", try, duration)
 					flaky++
+					testFlaked(test.name, try)
 				}
 				break
 			}
@@ -307,12 +401,28 @@ func main() {
 	}
 
 	// Clean up temp dir.
-	log.Printf("Removing temp dir %v", tmpDir)
-	if err := os.RemoveAll(tmpDir); err != nil {
-		log.Printf("Failed to remove temp dir: %v", err)
+	if tmpDir != "" {
+		log.Printf("Removing temp dir %v", tmpDir)
+		if err := os.RemoveAll(tmpDir); err != nil {
+			log.Printf("Failed to remove temp dir: %v", err)
+		}
 	}
 
-	// Print stats.
+	// Print summary.
+	log.Printf(strings.Repeat("=", 50))
+	for _, t := range tests {
+		switch {
+		case t.pass > 0 && t.fail == 0:
+			log.Printf("%-32s\tPASS", t.name)
+		case t.pass > 0 && t.fail > 0:
+			log.Printf("%-32s\tFLAKY (%v/%v failed)", t.name, t.fail, t.pass+t.fail)
+		case t.pass == 0 && t.fail > 0:
+			log.Printf("%-32s\tFAIL (%v tries)", t.name, t.fail)
+		case t.pass == 0 && t.fail == 0:
+			log.Printf("%-32s\tSKIPPED", t.name)
+		}
+	}
+	log.Printf(strings.Repeat("=", 50))
 	skipped := len(tests) - passed - flaky - failed
 	log.Printf("%v PASSED, %v FLAKY, %v FAILED, %v SKIPPED", passed, flaky, failed, skipped)
 	log.Printf("Total time: %v", time.Since(startTime))
@@ -320,4 +430,255 @@ func main() {
 	if failed > 0 || skipped > 0 {
 		os.Exit(1)
 	}
+}
+
+func updateEnv(orig []string, updates map[string]string) []string {
+	var env []string
+	for _, v := range orig {
+		parts := strings.SplitN(v, "=", 2)
+		if _, ok := updates[parts[0]]; !ok {
+			env = append(env, v)
+		}
+	}
+	for k, v := range updates {
+		env = append(env, k+"="+v)
+	}
+	return env
+}
+
+type Stats struct {
+	TestStats map[string]TestStats
+}
+
+type TestStats struct {
+	Pass, Fail, Flake int
+	PassTime          time.Duration
+
+	name string
+}
+
+func sendStats(values url.Values) {
+	if *remoteStats != "" {
+		log.Printf("Sending remote stats to %v", *remoteStats)
+		if _, err := http.PostForm(*remoteStats, values); err != nil {
+			log.Printf("Can't send remote stats: %v", err)
+		}
+	}
+}
+
+func testPassed(name string, passTime time.Duration) {
+	sendStats(url.Values{
+		"test":     {name},
+		"result":   {"pass"},
+		"duration": {passTime.String()},
+	})
+	updateTestStats(name, func(ts *TestStats) {
+		totalTime := int64(ts.PassTime)*int64(ts.Pass) + int64(passTime)
+		ts.Pass++
+		ts.PassTime = time.Duration(totalTime / int64(ts.Pass))
+	})
+}
+
+func testFailed(name string) {
+	sendStats(url.Values{
+		"test":   {name},
+		"result": {"fail"},
+	})
+	updateTestStats(name, func(ts *TestStats) {
+		ts.Fail++
+	})
+}
+
+func testFlaked(name string, try int) {
+	sendStats(url.Values{
+		"test":   {name},
+		"result": {"flake"},
+		"try":    {strconv.FormatInt(int64(try), 10)},
+	})
+	updateTestStats(name, func(ts *TestStats) {
+		ts.Flake += try - 1
+	})
+}
+
+func updateTestStats(name string, update func(*TestStats)) {
+	var stats Stats
+
+	data, err := ioutil.ReadFile(statsFileName)
+	if err != nil {
+		log.Print("Can't read stats file, starting new one.")
+	} else {
+		if err := json.Unmarshal(data, &stats); err != nil {
+			log.Printf("Can't parse stats file: %v", err)
+			return
+		}
+	}
+
+	if stats.TestStats == nil {
+		stats.TestStats = make(map[string]TestStats)
+	}
+	ts := stats.TestStats[name]
+	update(&ts)
+	stats.TestStats[name] = ts
+
+	data, err = json.MarshalIndent(stats, "", "\t")
+	if err != nil {
+		log.Printf("Can't encode stats file: %v", err)
+		return
+	}
+	if err := ioutil.WriteFile(statsFileName, data, 0644); err != nil {
+		log.Printf("Can't write stats file: %v", err)
+	}
+}
+
+func reshardTests(config *Config, numShards int) error {
+	var stats Stats
+
+	var data []byte
+	if *remoteStats != "" {
+		log.Printf("Using remote stats for resharding: %v", *remoteStats)
+		resp, err := http.Get(*remoteStats)
+		if err != nil {
+			return err
+		}
+		if data, err = ioutil.ReadAll(resp.Body); err != nil {
+			return err
+		}
+	} else {
+		var err error
+		data, err = ioutil.ReadFile(statsFileName)
+		if err != nil {
+			return errors.New("can't read stats file")
+		}
+	}
+
+	if err := json.Unmarshal(data, &stats); err != nil {
+		return fmt.Errorf("can't parse stats file: %v", err)
+	}
+
+	// Sort tests by PassTime.
+	var tests []TestStats
+	var totalTime int64
+	for name, test := range stats.TestStats {
+		test.name = name
+		tests = append(tests, test)
+		totalTime += int64(test.PassTime)
+	}
+	sort.Sort(ByPassTime(tests))
+
+	// Group into shards.
+	max := totalTime / int64(numShards)
+	shards := make([][]TestStats, numShards)
+	sums := make([]int64, numShards)
+	// First pass: greedy approximation.
+firstPass:
+	for len(tests) > 0 {
+		v := int64(tests[0].PassTime)
+
+		for n := range shards {
+			if sums[n]+v < max {
+				shards[n] = append(shards[n], tests[0])
+				sums[n] += v
+				tests = tests[1:]
+				continue firstPass
+			}
+		}
+		// None of the bins has room. Go to second pass.
+		break
+	}
+	// Second pass: distribute the remainder.
+	for len(tests) > 0 {
+		nmin := 0
+		min := sums[0]
+
+		for n := range sums {
+			if sums[n] < min {
+				nmin = n
+				min = sums[n]
+			}
+		}
+
+		shards[nmin] = append(shards[nmin], tests[0])
+		sums[nmin] += int64(tests[0].PassTime)
+		tests = tests[1:]
+	}
+
+	// Update config and print results.
+	for i, tests := range shards {
+		for _, t := range tests {
+			config.Tests[t.name].Shard = i
+			log.Printf("% 32v:\t%v\n", t.name, t.PassTime)
+		}
+		log.Printf("Shard %v total: %v\n", i, time.Duration(sums[i]))
+	}
+
+	return nil
+}
+
+type ByPassTime []TestStats
+
+func (a ByPassTime) Len() int           { return len(a) }
+func (a ByPassTime) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
+func (a ByPassTime) Less(i, j int) bool { return a[i].PassTime > a[j].PassTime }
+
+func getTestsSorted(names []string, testMap map[string]*Test) []*Test {
+	sort.Strings(names)
+	var tests []*Test
+	for _, name := range names {
+		t := testMap[name]
+		t.name = name
+		tests = append(tests, t)
+	}
+	return tests
+}
+
+func selectedTests(config *Config) []*Test {
+	var tests []*Test
+	if *shard >= 0 {
+		// Run the tests in a given shard.
+		// This can be combined with positional args.
+		var names []string
+		for name, t := range config.Tests {
+			if t.Shard == *shard {
+				t.name = name
+				names = append(names, name)
+			}
+		}
+		tests = getTestsSorted(names, config.Tests)
+	}
+	if flag.NArg() > 0 {
+		// Positional args for manual selection.
+		for _, name := range flag.Args() {
+			t, ok := config.Tests[name]
+			if !ok {
+				log.Fatalf("Unknown test: %v", name)
+			}
+			t.name = name
+			tests = append(tests, t)
+		}
+	}
+	if flag.NArg() == 0 && *shard < 0 {
+		// Run all tests.
+		var names []string
+		for name := range config.Tests {
+			if !config.Tests[name].Manual {
+				names = append(names, name)
+			}
+		}
+		tests = getTestsSorted(names, config.Tests)
+	}
+	return tests
+}
+
+var (
+	port      = 16000
+	portMutex sync.Mutex
+)
+
+func getPortStart(size int) int {
+	portMutex.Lock()
+	defer portMutex.Unlock()
+
+	start := port
+	port += size
+	return start
 }
